+++ src/prefabs/structures/Appbar/options/index.ts
@@ -32,7 +32,7 @@ export const categories = [
     ],
   },
   {
-    label: 'Advanced settings',
+    label: 'Advanced Options',
     expanded: false,
     members: ['dataComponentAttribute'],
   },