+++ src/prefabs/structures/ListSubHeader/options/index.ts
@@ -8,7 +8,7 @@ import { advanced } from '../../advanced';
 
 export const categories = [
   {
-    label: 'Advanced settings',
+    label: 'Advanced Options',
     expanded: false,
     members: ['dataComponentAttribute'],
   },